Add component test
Add test for the classmentors component and for the ace of coders component.

I picked ace of coder because it’s very small. I converted it to an Angular 1.5 component and moved fetching the stats to a component attribute resolved during routing.

Fetching the stats is now a service that return a promise: it’s easy to test on its own. The url to fetch is an angular constant; we do not need to hard code the url to fetch in the test (maybe it doesn’t need to be an angular constant, just an exported value).

TODO: add tests for other components